new file:   beautifulsoup/beautifuldemo.py
	new file:   beautifulsoup/soupmmscrapy.py
	new file:   ichatrobot/ichatrobot.py
	new file:   selenium/aisinei/taobaobraspider.py
	modified:   selenium/seleniumcookies2.py
	modified:   selenium/tmbra/taobaobraspider.py

+++ selenium/tmbra/taobaobraspider.py
@@ -16,20 +16,22 @@ import os
 import sys
 class SeleniumCookie(object):
     def __init__(self,url):
-        options = webdriver.ChromeOptions()
+        option = webdriver.ChromeOptions()
         # 设置中文
-        options.add_argument('lang=zh_CN.UTF-8')
+        option.add_argument('lang=zh_CN.UTF-8')
         #options.add_argument(r"user-data-dir=C:\Users\secondtonone1\AppData\Local\Google\Chrome\User Data")
         self.url_=url
-        self.driver_ = webdriver.Chrome(options=options)
+        self.driver_ = webdriver.Chrome(chrome_options=option)
         self.driver_.get(self.url_)
         self.path=os.path.dirname(os.path.abspath(__file__))
         self.cookiepath = os.path.join(self.path,"seleniumcookies.pk1")
-        #self.save_cookie()
+        self.save_cookie()
         self.load_cookie()
         self.refresh_page()
         self.wait = WebDriverWait(self.driver_,10)
     def save_cookie(self):
+        if(os.path.exists(self.cookiepath)):
+            return
         time.sleep(60)
         with open (self.cookiepath,"wb") as f:
             pickle.dump(self.driver_.get_cookies(),f)
@@ -105,7 +107,7 @@ class SeleniumCookie(object):
        
        
 if __name__ == "__main__":
-    seleniumcookie = SeleniumCookie('https://www.tmall.com/')
+    seleniumcookie = SeleniumCookie('https://www.taobao.com/')
     seleniumcookie.openWindow('https://detail.tmall.com/item.htm?spm=a220m.1000858.1000725.1.2e0d63ffvOPH2N&id=575198548137&skuId=3774938064975&areaId=110100&user_id=1644123097&cat_id=2&is_b=1&rn=a2781533c3ad59ab4c24d1f4246113b2')
     seleniumcookie.closeDialog()
     seleniumcookie.clickComment()
